feat: add LocatorSuggestProvider to sherlock
Includes a refactor to pull providers out into a separate file. Shouldn't break any existing apps.
