+++ .vscode/settings.json
@@ -2,6 +2,7 @@
   "cSpell.words": [
     "Centerlines",
     "SGID",
+    "SRID",
     "TSQL",
     "Topo",
     "agrc",
