+++ packages/sherlock/package.json
@@ -8,7 +8,7 @@
     "type": "git",
     "url": "https://github.com/agrc/kitchen-sink.git"
   },
-  "main": "./dist/Sherlock.js",
+  "main": "./dist/index.js",
   "files": [
     "/dist"
   ],
