+++ packages/sherlock/src/index.js
@@ -1,0 +1,5 @@
+import Sherlock from './Sherlock';
+import { WebApiProvider, MapServiceProvider, LocatorSuggestProvider } from './providers';
+
+
+export { Sherlock, WebApiProvider, MapServiceProvider, LocatorSuggestProvider };
